requirements.txt
gunicorn==20.0.4
flask==2.0.3
pillow==9.0.1
tensorflow==1.14.0
numpy==1.21.5